Rename Mamba to Hybrid outside megatron/core (#4159)
