+++ examples/post_training/modelopt/conf/nvidia/Nemotron-H-4B-Instruct.sh
@@ -38,5 +38,5 @@ MODEL_ARGS=" \
     --make-vocab-size-divisible-by 1 \
     --use-mcore-models \
     --rotary-base 10000 \
-    --export-model-type MambaModel \
+    --export-model-type HybridModel \
 "
