+++ examples/post_training/modelopt/conf/nvidia/Nemotron-H-56B-Base-8K.sh
@@ -35,5 +35,5 @@ MODEL_ARGS=" \
     --max-position-embeddings 8192 \
     --tokenizer-type HuggingFaceTokenizer \
     --bf16 \
-    --export-model-type MambaModel \
+    --export-model-type HybridModel \
     "
