+++ examples/post_training/modelopt/conf/nvidia/Nemotron-H-8B-Base-8K.sh
@@ -37,6 +37,6 @@ MODEL_ARGS=" \
     --use-mcore-models \
     --rotary-percent 0.5 \
     --rotary-base 500000 \
-    --export-model-type MambaModel \
+    --export-model-type HybridModel \
 "
 #    --rotary-base 10000 \
