+++ megatron/post_training/arguments.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2024, NVIDIA CORPORATION. All rights reserved.
+# Copyright (c) 2024-2026, NVIDIA CORPORATION. All rights reserved.
 
 
 def add_modelopt_args(parser):
@@ -10,8 +10,9 @@ def add_modelopt_args(parser):
         "--export-model-type",
         type=str,
         default="GPTModel",
-        choices=["GPTModel", "MambaModel"],
-        help="Model type to use in model_provider.",
+        choices=["GPTModel", "HybridModel", "MambaModel"],
+        help='Model type to use in model_provider. Use "HybridModel" for hybrid models '
+        '(formerly MambaModel). "MambaModel" is accepted for backward compatibility but deprecated.',
     )
     group.add_argument(
         "--export-legacy-megatron",
