+++ model_provider.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2025, NVIDIA CORPORATION.  All rights reserved.
+# Copyright (c) 2025-2026, NVIDIA CORPORATION.  All rights reserved.
 
 """Common functions used in train_*.py and pretrain_*.py scripts."""
 
@@ -7,11 +7,11 @@ from typing import Callable, Optional, Union
 import torch
 
 from megatron.core.models.gpt import GPTModel
-from megatron.core.models.mamba import MambaModel
+from megatron.core.models.hybrid.hybrid_model import HybridModel
 from megatron.training import get_args, print_rank_0
 
 try:
-    from megatron.post_training.model_builder import modelopt_gpt_mamba_builder
+    from megatron.post_training.model_builder import modelopt_gpt_hybrid_builder
     has_nvidia_modelopt = True
 except ImportError:
     has_nvidia_modelopt = False
@@ -23,18 +23,18 @@ import megatron.legacy.model  # isort: skip
 
 def model_provider(
     model_builder: Callable, pre_process=True, post_process=True, vp_stage: Optional[int] = None, config=None, pg_collection=None,
-) -> Union[GPTModel, megatron.legacy.model.GPTModel, MambaModel]:
+) -> Union[GPTModel, megatron.legacy.model.GPTModel, HybridModel]:
     """Builds the model.
 
     If you set the use_legacy_models to True, it will return the legacy GPT model and if not the mcore GPT model.
 
     Args:
-        model_builder: A callable that builds the actual model, its signature is the same as model_provider's with an exception of the first argument which is a builder itself. In addition might take a config passed from outside to skip its own config loading. See gpt_builder or mamba_builder for an example, see _gpt_model_builder in train_rl.py to see how to augment a default gpt builder and pass the config from outside
+        model_builder: A callable that builds the actual model, its signature is the same as model_provider's with an exception of the first argument which is a builder itself. In addition might take a config passed from outside to skip its own config loading. See gpt_builder or hybrid_builder for an example, see _gpt_model_builder in train_rl.py to see how to augment a default gpt builder and pass the config from outside
         pre_process (bool, optional): Set to true if you need to compute embedings. Defaults to True.
         post_process (bool, optional): Set to true if you need to compute output logits/loss. Defaults to True.
 
     Returns:
-        Union[GPTModel, megatron.legacy.model.GPTModel, MambaModel]: The returned model
+        Union[GPTModel, megatron.legacy.model.GPTModel, HybridModel]: The returned model
     """
     args = get_args()
 
@@ -58,7 +58,7 @@ def model_provider(
 
     if has_nvidia_modelopt and getattr(args, 'modelopt_enabled', False):
         # [ModelOpt]: Use custom builder + spec when modelopt is enabled
-        model_builder = modelopt_gpt_mamba_builder
+        model_builder = modelopt_gpt_hybrid_builder
 
     return model_builder(args, pre_process, post_process, vp_stage, config=config, pg_collection=pg_collection)
