+++ tools/run_text_generation_server.py
@@ -15,7 +15,7 @@ from contextlib import nullcontext
 import torch
 
 from gpt_builders import gpt_builder
-from mamba_builders import mamba_builder
+from hybrid_builders import hybrid_builder
 from megatron.core.inference.contexts import StaticInferenceContext
 from megatron.core.inference.engines import AbstractEngine, StaticInferenceEngine
 from megatron.core.inference.engines.abstract_engine import AbstractEngine
@@ -140,8 +140,16 @@ def main(model_type: str = "gpt"):
         # Set up model and load checkpoint
         if model_type == "gpt":
             model_builder = gpt_builder
-        elif model_type == "mamba":
-            model_builder = mamba_builder
+        elif model_type in ("hybrid", "mamba"):
+            if model_type == "mamba":
+                import warnings
+
+                warnings.warn(
+                    'model_type="mamba" is deprecated. Use model_type="hybrid" instead.',
+                    DeprecationWarning,
+                    stacklevel=2,
+                )
+            model_builder = hybrid_builder
         else:
             raise ValueError(f"Invalid model provider {model_type}")
         model = get_model(partial(model_provider, model_builder), wrap_with_ddp=False)
